DGA Items, Dialog, and Folder Restructure
Restructured the folder to add in DGA recipe items for mail. Also added dialog to include with animations, an animations file that the content.json should look up (no animations made yet), and the Gossip file which is for commentary made by other NPCs in regards to Antoine's events. Only intro event included so far.